Move BackButton to common directory
