+++ assets/js/components/ClusterSettingsPage/ClusterSettingsPage.jsx
@@ -18,7 +18,7 @@ import { executionRequested } from '@state/lastExecutions';
 import { UNKNOWN_PROVIDER, VMWARE_PROVIDER, TARGET_CLUSTER } from '@lib/model';
 
 import PageHeader from '@components/PageHeader';
-import BackButton from '@components/BackButton';
+import BackButton from '@common/BackButton';
 
 import { ClusterInfoBox } from '@pages/ClusterDetails';
 import LoadingBox from '@components/LoadingBox';
