+++ assets/js/components/ExecutionResults/CheckResultDetail/BackToTargetExecution.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { TARGET_CLUSTER, TARGET_HOST } from '@lib/model';
 
-import BackButton from '@components/BackButton';
+import BackButton from '@common/BackButton';
 
 function BackToTargetExecution({ targetID, targetType }) {
   const targetTypeToExecutionURL = {
