+++ assets/js/components/HostDetails/HostDetails.jsx
@@ -7,7 +7,7 @@ import { agentVersionWarning } from '@lib/agent';
 import Button from '@common/Button';
 import Table from '@common/Table';
 import PageHeader from '@components/PageHeader';
-import BackButton from '@components/BackButton';
+import BackButton from '@common/BackButton';
 
 import CleanUpButton from '@components/CleanUpButton';
 import DeregistrationModal from '@components/DeregistrationModal';
